+++ python/GafferRenderManTest/RenderPassAdaptorTest.py
@@ -52,12 +52,6 @@ class RenderPassAdaptorTest( GafferSceneTest.RenderPassAdaptorTest ) :
 	shadowColor = imath.Color4f( 1, 1, 1, 0 )
 	litColor = imath.Color4f( 0 )
 
-	## \todo Remove once light linking is supported.
-	@unittest.skip( "Light linking not supported yet" )
-	def testReflectionCasterLightLinks( self ) :
-
-		pass
-
 	def _createDistantLight( self ) :
 
 		light = GafferRenderMan.RenderManLight()