+++ Changes.md
@@ -4,6 +4,7 @@
 Improvements
 ------------
 
+- RenderMan : Added light linking support.
 - VisualiserTool : Added new visualisation for orientation (Quatf) data.
 - PrimitiveInspector : Changed column order for quaternions to match Imath's conventions.